MIPS architecture operation codes simulator V1
